Image Processing Files using OpenCV
A series of frameworks/examples to work with that implement all the
wanted features of image processing. Runs on Python2 after installing
the necessary packages, which are imported in practically every file.
The highlight of this is that through all the tools available we can
then use everything to fine tune image detection through the xml files
that outline human features.