Experiments
Took all old files from YouTube tutorials and put them in a new Folder
titled BasicTutorials.

All new implemented files are within the Experiments folder. The main
files worth attention is the CornerDetection, the
live_CameraThresholding (this is the most prevalent program for
experimenting filters), and the pixel_intColorRange (this shows the
integer values for pixels with respect to the BGR values per filter
appliance).