Blur Detection
Added an additional feature to the main building
liveCamera_thresholding so now aside from reading frames in from
threads to increase FPS, we also added blur detection on a grayscaled
form of the frame.
This will add to the feature of the drone being able to understand
which frames it’s reading are blurry so it can discard sending those to
mother station for meshing a 3d model.